optimize

+++ src/Drop.java
@@ -13,28 +13,28 @@ public class Drop {
         i2Treshold = Math.round(0.7f * (i2Max-1));
     }
 
-    public double p1(int i1) {
-        double p;
+    public float p1(int i1) {
+        float p;
         if (i1 < i1Treshold) {
             p = 0;
         } else if (i1 >= i1Max) {
             p = 1;
         } else {
-            p = (i1 - i1Treshold)*1.0/ (i1Max - i1Treshold);
+            p = (i1 - i1Treshold)*1.0f/ (i1Max - i1Treshold);
         }
 //        System.out.printf("p= %f, treshold= %d, iMax= %d, i= %d\n", p, i1Treshold, i1Max, i1);
         return p;
     }
 
-    public double p2(int i2) {
+    public float p2(int i2) {
         if (i2 < i2Treshold) return 0;
         if (i2 >= i2Max) return 1;
-        return (i2 - i2Treshold)*1.0/(i2Max - i2Treshold);
+        return (i2 - i2Treshold)*1.0f/(i2Max - i2Treshold);
     }
 
-    public double percent(double[] probabilities) {
+    public float percent(float[] probabilities) {
 
-        double sum = 0;
+        float sum = 0;
         for (int i1 = 0; i1 <= i1Max; i1++) {
             for (int i2 = 0; i2 <= i2Max; i2++) {
                 sum += probabilities[i1*(i2Max+1)+i2]*(p1(i1) + p2(i2));
@@ -43,9 +43,9 @@ public class Drop {
         return sum;
     }
 
-    public double percentD1(double[] probabilities) {
+    public float percentD1(float[] probabilities) {
 
-        double sum = 0;
+        float sum = 0;
         for (int i1 = 0; i1 <= i1Max; i1++) {
             for (int i2 = 0; i2 <= i2Max; i2++) {
                 sum += probabilities[i1*(i2Max+1)+i2]*(p1(i1));
@@ -54,9 +54,9 @@ public class Drop {
         return sum;
     }
 
-    public double percentD2(double[] probabilities) {
+    public float percentD2(float[] probabilities) {
 
-        double sum = 0;
+        float sum = 0;
         for (int i1 = 0; i1 <= i1Max; i1++) {
             for (int i2 = 0; i2 <= i2Max; i2++) {
                 sum += probabilities[i1*(i2Max+1)+i2]*(p2(i2));
